censys-tenable-vm/requirements.txt
# install these using 'pip install --upgrade -r requirements.txt' 

censys
pytenable